reduce code duplication in SeeAll forms
