+++ imports/framework/Forms/DetailsForm/DetailsForm.Helpers.js
@@ -1,6 +1,7 @@
 import { Meteor } from 'meteor/meteor';
 import { Template } from 'meteor/templating';
 import { FlowRouter } from 'meteor/kadira:flow-router';
+import { TAPi18n } from 'meteor/tap:i18n';
 
 import RoleManager from '/imports/framework/Managers/RoleManager';
 
@@ -71,4 +72,35 @@ function hasPermissionToSee(definition) {
   return hasRole && customFulfilled;
 }
 
-export { loadData, getValue, getKey, isType, hasPermissionToSee };
+function getValueForSeeAllItems(definition, entity) {
+  const array = getValue(definition, entity);
+  const moreItemsCount = array.length - definition.maxItemsShown;
+  let messagePathParts = FlowRouter.getRouteName().split('.');
+
+  messagePathParts.pop();
+  messagePathParts.splice(1, 0, 'entity');
+  messagePathParts.push(definition.key);
+  messagePathParts.push('seeAllItems');
+
+  return TAPi18n.__(messagePathParts.join('.'), { count: moreItemsCount });
+}
+
+function clickHandlerForSeeAllItems(e) {
+  e.stopPropagation();
+
+  const data = Template.currentData();
+  let params = FlowRouter.current().params;
+  const route = data.definition.allItemsRoute;
+
+  FlowRouter.go(FlowRouter.path(route, params));
+}
+
+export {
+  loadData,
+  getValue,
+  getKey,
+  isType,
+  hasPermissionToSee,
+  getValueForSeeAllItems,
+  clickHandlerForSeeAllItems
+};
